+++ apps/knockdog/src/widgets/bottom-nav-bar/config/navitem.ts
@@ -1,0 +1,6 @@
+export const NAV_ITEMS = [
+  { href: '/', icon: 'SearchNav' as const, label: '내 주변' },
+  { href: '/save', icon: 'SaveNav' as const, label: '저장' },
+  { href: '/compare', icon: 'CompareNav' as const, label: '비교' },
+  { href: '/my', icon: 'MypageNav' as const, label: '마이' },
+];
