+++ apps/knockdog/src/widgets/bottom-nav-bar/index.ts
@@ -1,1 +1,1 @@
-export * from './BottomNavBar';
+export * from './ui/BottomNavBar';
